requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
pyodbc==4.0.39
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
plotly==5.17.0
Jinja2==3.1.2